[WIP] superio/ite: Add support for IT8625E
We don't have datasheet for this chip, but ITE doesn't change their
design much.

Found on tiny firewall appliance from Chinese company named "Topton"
with Intel N100 SoC. Because this system is fanless, all we really
need is the ability to use cisco-style serial console on front of
the device, which is working.

Change-Id: I9c27f52785d294a6f7c315b8df47d4dd5b389414
Signed-off-by: Alicja Michalska <ahplka19@gmail.com>